Fix PLIC interrupt delivery, Altera avail fix
PLIC interrupt delivery was broken because it didn't cleared
the INTERRUPT_SEXTERNAL bit in mip register. This made the
interrupt always-on, which is why the kernel was hanging.

Altera controller was buggy because it always lost one byte -
the RAVAIL number includes the byte being read.

Also contains yet another coordinate report fix :)