Dual-license under GPL 3.0 OR MPL 2.0
This change allows to use this work as a component of a larger, non-GPL compliant project in any form (Static linkage, binary distribution, modules) as long as the receipts comply with the MPL 2.0 license.
Most notable use cases arising from this change:
- Linking to librvvm statically, without any implications on the resulting software license as a whole. Permissive stays permissive, proprietary stays proprietary.
- Distributing RVVM in environments that partially violate third law of the GPL (The freedom to modify). This includes Apple App Store (You may modify the source code, but may not run it) or BSD system repositories (The entire system may be forked as closed source).

For any form of software distribution, changes made precisely to this project should be made public to comply with either licenses.
This form of weak copyleft should prevent proliferation of proprietary forks, and useful changes be considered for upstreaming.

As for why MPL and not LGPL: LGPL view on the "derived work" or "linkage" is blurred, and the end-users may accidentally violate it.
Dual-licensing approach is questionable, but at this point dropping the GPL isn't going to bring any benefits. Discussions regarding this are welcome.

The (new) source license headers may transition to a dual-license header similar to the one below:

This Source Code Form is subject to the terms of the Mozilla Public
License, v. 2.0. If a copy of the MPL was not distributed with this
file, You can obtain one at https://mozilla.org/MPL/2.0/.

Alternatively, the contents of this file may be used under the terms
of the GNU General Public License as published by the Free Software
Foundation, either version 3 of the License, or any later version.

You should have received a copy of the GNU General Public License
along with this program.  If not, see <https://www.gnu.org/licenses/>.